+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement class-based handler execution so that we can perform `this.<handler>(props)` from inside other handlers and access and update config options like `messages.error` or `maxFileSize` etc.
 
+### Fixed
+
+- Replace `Blob` instance types with an empty object during path extraction so that we don't get paths like `file.lastModified` or `file.stream`
+
 ## [1.1.9] - 2025-06-25
 
 ### Added